refactor: 테스트 패키지 분리
미션 진행을 원할하게 하기 위해
로그인 기능 테스트와 지하철 기능 테스트를 분리시켰다.
