+++ src/test/java/nextstep/subway/acceptance/subway/LineSteps.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.subway;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
